refactor: move test generator util function to correct file
