+++ src/tests/generators/generateConversation.ts
@@ -151,51 +151,3 @@ export const populateConversationInEmailStore = ({
 	updateConversations([generatedConversation]);
 	return { conversation: generatedConversation, messages: generatedMessages };
 };
-
-/**
- * Populates the email store with messages and returns the generated messages.
- * The function generates messages based on provided message IDs or message generation parameters.
- * If neither is provided, it generates a default set of messages.
- *
- * It then updates the email store with the generated messages and returns them.
- */
-export const populateMessagesInEmailStore = ({
-	messageGeneratorParams,
-	messageIds,
-	messagesNumber = 1
-}: {
-	messageGeneratorParams?: Array<MessageGenerationParams>;
-	messageIds?: Array<string>;
-	messagesNumber?: number;
-}): Array<MailMessage> => {
-	// Generate messages based on provided message IDs
-	const messagesFromMessageIds = messageIds?.map((messageId) =>
-		generateMessage({
-			id: messageId,
-			folderId: FOLDERS.INBOX,
-			cid: '1'
-		})
-	);
-
-	// Generate messages based on provided message generation parameters
-	const messagesFromMessageGeneratorParams = messageGeneratorParams?.map((messageGeneratorParam) =>
-		generateMessage({ ...messageGeneratorParam })
-	);
-
-	// Generate default messages if no message IDs or parameters are provided
-	const defaultMessages = Array.from({ length: messagesNumber }).map((_, index) =>
-		generateMessage({
-			id: (index + 100).toString(),
-			folderId: FOLDERS.INBOX,
-			cid: '1'
-		})
-	);
-
-	// Use the provided messages or fall back to default messages
-	const generatedMessages =
-		messagesFromMessageIds ?? messagesFromMessageGeneratorParams ?? defaultMessages;
-
-	updateMessages(generatedMessages);
-
-	return generatedMessages;
-};
