+++ src/tests/generators/tests/generateMessage.test.ts
@@ -4,7 +4,7 @@
  * SPDX-License-Identifier: AGPL-3.0-only
  */
 
-import { populateMessagesInEmailStore } from '../generateConversation';
+import { populateMessagesInEmailStore } from '../generateMessage';
 
 describe('generateConversation', () => {
 	describe('populateMessagesInEmailStore', () => {
